+++ fs/osfs.py
@@ -144,6 +144,7 @@ class OSFS(FS):
             "network": False,
             "read_only": False,
             "supports_rename": True,
+            "supports_mtime": False,
             "thread_safe": True,
             "unicode_paths": os.path.supports_unicode_filenames,
             "virtual": False,